Add tests to cover main.go
