+++ layli.go
@@ -1,12 +1,9 @@
 package layli
 
 import (
-	"fmt"
-	"io"
 	"strings"
 
 	svg "github.com/ajstarks/svgo"
-	"gopkg.in/yaml.v3"
 )
 
 type OutputFunc func(output string) error
@@ -16,30 +13,6 @@ type Diagram struct {
 	Config   Config
 	Layout   *Layout
 	ShowGrid bool
-}
-
-func NewConfigFromFile(r io.Reader) (*Config, error) {
-	config := Config{}
-	err := yaml.NewDecoder(r).Decode(&config)
-	if err != nil {
-		return nil, fmt.Errorf("reading config file: %w", err)
-	}
-	config.Spacing = 20
-
-	if config.NodeWidth == 0 {
-		config.NodeWidth = 5
-	}
-	if config.NodeHeight == 0 {
-		config.NodeHeight = 3
-	}
-	if config.Margin == 0 {
-		config.Margin = 2
-	}
-	if config.Border == 0 {
-		config.Border = 1
-	}
-
-	return &config, nil
 }
 
 // Draw turns the diagram in to an image
