+++ test/fixtures/inputs/bad-config.layli
@@ -1,0 +1,4 @@
+nodes:
+  - id: 
+
+{
